feat(cli): verifyEntry guard on npm store-hit + commit propagation
Wrap the npm store-hit short-circuit in install.ts around line 274
with a verifyEntry() check. A store entry whose stamp is missing or
whose content hash does not match is moved to <askHome>/.quarantine/
and the install falls through to a fresh fetch from the source
adapter. The same quarantineEntry helper is used by the github source
store-hit path (T005) — it moves to the shared store/index.ts.

Propagate result.meta.commit into ResolvedEntry.commit so github
installs record the 40-char hex SHA the tag resolved to at install
time. For npm/web/llms-txt meta.commit is unset, so the field stays
undefined.

Github source switches from its private quarantineEntry to the shared
one now exposed from store/index.ts.

End-to-end quarantine + commit propagation tests land in T010 as
integration scenarios against the full install pipeline.

Part of T006 for track github-store-pm-unified-20260411.
Refs #68
